requirements.txt
ipykernel
wheel
scipy
numpy<1.22
pandas
scikit-learn
matplotlib
seaborn
opentsne==0.6.2
umap-learn==0.5.3